feat: 불필요 인터페이스 삭제
