+++ src/main/java/chess/WebApplication.java
@@ -6,7 +6,7 @@ import static spark.Spark.port;
 import static spark.Spark.staticFileLocation;
 
 import chess.dao.MemoryPieceDaoImpl;
-import chess.service.ChessServiceImpl;
+import chess.service.ChessService;
 import chess.web.controller.ChessWebController;
 import spark.template.handlebars.HandlebarsTemplateEngine;
 
@@ -18,7 +18,7 @@ public class WebApplication {
         port(8080);
         staticFileLocation("/static");
 
-        ChessWebController chessWebController = new ChessWebController(new ChessServiceImpl(new MemoryPieceDaoImpl()));
+        ChessWebController chessWebController = new ChessWebController(new ChessService(new MemoryPieceDaoImpl()));
 
         get("/", chessWebController::index, HANDLEBARS_TEMPLATE_ENGINE);
         get("/start", chessWebController::create, HANDLEBARS_TEMPLATE_ENGINE);
